+++ packages/radio-group/test/radio-button-lit.test.js
@@ -1,0 +1,2 @@
+import '../src/vaadin-lit-radio-button.js';
+import './radio-button.common.js';
