+++ packages/radio-group/test/radio-button-polymer.test.js
@@ -1,0 +1,2 @@
+import '../src/vaadin-radio-button.js';
+import './radio-button.common.js';
