+++ packages/radio-group/test/radio-button.common.js
@@ -2,7 +2,6 @@ import { expect } from '@esm-bundle/chai';
 import { fire, fixtureSync, mousedown, mouseup, nextFrame } from '@vaadin/testing-helpers';
 import { resetMouse, sendKeys, sendMouse } from '@web/test-runner-commands';
 import sinon from 'sinon';
-import '../vaadin-radio-button.js';
 
 describe('radio-button', () => {
   let radio, input, label;
@@ -25,8 +24,9 @@ describe('radio-button', () => {
   });
 
   describe('default', () => {
-    beforeEach(() => {
+    beforeEach(async () => {
       radio = fixtureSync('<vaadin-radio-button label="Label"></vaadin-radio-button>');
+      await nextFrame();
       label = radio.querySelector('[slot=label]');
     });
 
@@ -44,8 +44,9 @@ describe('radio-button', () => {
   });
 
   describe('native input', () => {
-    beforeEach(() => {
+    beforeEach(async () => {
       radio = fixtureSync('<vaadin-radio-button></vaadin-radio-button>');
+      await nextFrame();
       input = radio.querySelector('[slot=input]');
       label = radio.querySelector('[slot=label]');
     });
@@ -110,8 +111,9 @@ describe('radio-button', () => {
   });
 
   describe('disabled attribute', () => {
-    beforeEach(() => {
+    beforeEach(async () => {
       radio = fixtureSync('<vaadin-radio-button disabled></vaadin-radio-button>');
+      await nextFrame();
       input = radio.querySelector('[slot=input]');
       label = radio.querySelector('[slot=label]');
     });
@@ -138,8 +140,9 @@ describe('radio-button', () => {
 
   // TODO: A legacy suit. Replace with snapshot tests when possible.
   describe('active attribute', () => {
-    beforeEach(() => {
+    beforeEach(async () => {
       radio = fixtureSync('<vaadin-radio-button></vaadin-radio-button>');
+      await nextFrame();
       input = radio.querySelector('[slot=input]');
     });
 
@@ -167,8 +170,9 @@ describe('radio-button', () => {
   describe('change event', () => {
     let spy;
 
-    beforeEach(() => {
+    beforeEach(async () => {
       radio = fixtureSync('<vaadin-radio-button></vaadin-radio-button>');
+      await nextFrame();
       label = radio.querySelector('[slot=label]');
       input = radio.querySelector('[slot=input]');
 
@@ -195,19 +199,22 @@ describe('radio-button', () => {
       expect(spy.calledOnce).to.be.true;
     });
 
-    it('should not fire on programmatic toggle', () => {
+    it('should not fire on programmatic toggle', async () => {
       radio.checked = true;
+      await nextFrame();
       expect(spy.called).to.be.false;
     });
 
-    it('should not fire on input click when checked', () => {
+    it('should not fire on input click when checked', async () => {
       radio.checked = true;
+      await nextFrame();
       input.click();
       expect(spy.called).to.be.false;
     });
 
-    it('should not fire on label click when checked', () => {
+    it('should not fire on label click when checked', async () => {
       radio.checked = true;
+      await nextFrame();
       label.click();
       expect(spy.called).to.be.false;
     });
@@ -227,23 +234,26 @@ describe('radio-button', () => {
 
   describe('delegation', () => {
     describe('name attribute', () => {
-      beforeEach(() => {
-        radio = fixtureSync(`<vaadin-radio-button name="Name"></vaadin-radio-button>`);
+      beforeEach(async () => {
+        radio = fixtureSync('<vaadin-radio-button name="Name"></vaadin-radio-button>');
+        await nextFrame();
         input = radio.querySelector('[slot=input]');
       });
 
-      it('should delegate name attribute to the input', () => {
+      it('should delegate name attribute to the input', async () => {
         expect(input.getAttribute('name')).to.equal('Name');
 
         radio.removeAttribute('name');
+        await nextFrame();
         expect(input.hasAttribute('name')).to.be.false;
       });
     });
   });
 
   describe('has-label attribute', () => {
-    beforeEach(() => {
+    beforeEach(async () => {
       radio = fixtureSync('<vaadin-radio-button></vaadin-radio-button>');
+      await nextFrame();
     });
 
     it('should not set has-label attribute when label content is empty', () => {
